Backend/employee_tasks_saas/users/views.py
from django.shortcuts import render
from rest_framework import generics, permissions
from .serializers import UserSerializer, RegisterSerializer
from .models import CustomUser
from rest_framework.response import Response




# Create your views here.
class RegisterView(generics.CreateAPIView):
    queryset = CustomUser.objects.all()
    serializer_class = RegisterSerializer
    permission_classes = [permissions.AllowAny]



class UserDetailView(generics.RetrieveAPIView):
    queryset = CustomUser.objects.all()
    serializer_class = UserSerializer
    permission_classes = [permissions.IsAuthenticated]

    def get_object(self):
        return self.request.user

